rename: 재사용 컴포넌트 (dashboard)/_ui로 이동 (#119)
