+++ app/(dashboard)/_ui/details-information/invest-information.tsx
