+++ app/(dashboard)/_ui/details-information/percentage.tsx
