+++ app/(dashboard)/_ui/details-information/strategy-name-box.tsx
