+++ app/(dashboard)/_ui/details-information/styles.module.scss
