+++ app/(dashboard)/_ui/details-side-item/details-side-item.stories.tsx
