+++ app/(dashboard)/_ui/details-side-item/index.tsx
