+++ app/(dashboard)/_ui/details-side-item/styles.module.scss
